+++ test_app/www/js/test_options.js
@@ -1,0 +1,15 @@
+'use strict';
+
+function AdjustTestOptions() {
+    this.hasContext                       = false;
+    this.baseUrl                          = null;
+    this.gdprUrl                          = null;
+    this.basePath                         = null;
+    this.gdprPath                         = null;
+    this.useTestConnectionOptions         = null;
+    this.timerIntervalInMilliseconds      = null;
+    this.timerStartInMilliseconds         = null;
+    this.sessionIntervalInMilliseconds    = null;
+    this.subsessionIntervalInMilliseconds = null;
+    this.teardown                         = null;
+};