+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### Version 4.13.0 (15th May 2018)
+#### Added
+- Added `gdprForgetMe` method to `Adjust` interface to enable possibility for user to be forgotten in accordance with GDPR law.
+
+#### Native SDKs
+- [iOS@v4.13.0][ios_sdk_v4.13.0]
+- [Android@v4.13.0][android_sdk_v4.13.0]
+
+---
+
 ### Version 4.12.5 (12th March 2018)
 #### Native changes
 - Updated iOS SDK to `v4.12.3`.
@@ -359,6 +369,7 @@
 [ios_sdk_v4.11.5]: https://github.com/adjust/ios_sdk/tree/v4.11.5
 [ios_sdk_v4.12.1]: https://github.com/adjust/ios_sdk/tree/v4.12.1
 [ios_sdk_v4.12.3]: https://github.com/adjust/ios_sdk/tree/v4.12.3
+[ios_sdk_v4.13.0]: https://github.com/adjust/ios_sdk/tree/v4.13.0
 
 [android_sdk_v3.5.0]: https://github.com/adjust/android_sdk/tree/v3.5.0
 [android_sdk_v4.1.0]: https://github.com/adjust/android_sdk/tree/v4.1.0
@@ -374,3 +385,4 @@
 [android_sdk_v4.12.0]: https://github.com/adjust/android_sdk/tree/v4.12.0
 [android_sdk_v4.12.1]: https://github.com/adjust/android_sdk/tree/v4.12.1
 [android_sdk_v4.12.4]: https://github.com/adjust/android_sdk/tree/v4.12.4
+[android_sdk_v4.13.0]: https://github.com/adjust/android_sdk/tree/v4.13.0